Make Gc<'r, T> a newtype of &'r T
This makes it easier to pattern match on.
The down side is it commits Gc, to not being a AtomicPtr<T>.
Technically this is undefined behavior,
since the pointer may be overwritten by the Gc thread at any time.
Since the old pointer is valid for the duration of 'r,
we only rely on aligned writes of a word not being torn.
As a result the unsafe AtomicPtr is unneeded on ever ISA I'm aware of.